requirements.txt
mysql-connector-python
pymongo
python-dotenv
requests
pandas
 
 